refactor: Change the `user` app to the `accounts` app.
