+++ accounts/urls.py
@@ -1,12 +1,11 @@
 from django.urls import path
 
-from user import views
+from accounts import views
 
-app_name = 'user'
+app_name = 'accounts'
 
 urlpatterns = [
     path('create/', views.CreateUserView.as_view(), name='create'),
     path('token/', views.CreateTokenView.as_view(), name='token'),
     path('me', views.ManageUserView.as_view(), name='me')
-
 ]
